+++ dagster_university/dagster_and_dbt/dagster_and_dbt_tests/test_lesson_3.py
@@ -1,12 +1,16 @@
 import dagster as dg
-from dagster_and_dbt.lesson_3.assets import dbt
-from dagster_and_dbt.lesson_3.definitions import defs
-from dagster_and_dbt.lesson_3.resources import dbt_resource
+import pytest
 
-dbt_analytics_assets = dg.load_assets_from_modules(modules=[dbt])
+from dagster_and_dbt_tests.fixtures import setup_dbt_env  # noqa: F401
 
 
-def test_dbt_assets():
+@pytest.mark.parametrize("setup_dbt_env", ["lesson_3"], indirect=True)
+def test_dbt_assets(setup_dbt_env): # noqa: F811
+    from dagster_and_dbt.lesson_3.assets import dbt
+    from dagster_and_dbt.lesson_3.resources import dbt_resource
+
+    dbt_analytics_assets = dg.load_assets_from_modules(modules=[dbt])
+
     result = dg.materialize(
         assets=[*dbt_analytics_assets],
         resources={
@@ -16,5 +20,8 @@ def test_dbt_assets():
     assert result.success
 
 
-def test_def_can_load():
+@pytest.mark.parametrize("setup_dbt_env", ["lesson_3"], indirect=True)
+def test_def_can_load(setup_dbt_env): # noqa: F811
+    from dagster_and_dbt.lesson_3.definitions import defs
+
     assert defs
