+++ dagster_university/dagster_and_dbt/dagster_and_dbt_tests/test_lesson_6.py
@@ -1,12 +1,16 @@
 import dagster as dg
-from dagster_and_dbt.lesson_6.assets import dbt
-from dagster_and_dbt.lesson_6.definitions import defs
-from dagster_and_dbt.lesson_6.resources import dbt_resource
+import pytest
 
-dbt_analytics_assets = dg.load_assets_from_modules(modules=[dbt])
+from dagster_and_dbt_tests.fixtures import setup_dbt_env  # noqa: F401
 
 
-def test_dbt_partitioned_incremental_assets():
+@pytest.mark.parametrize("setup_dbt_env", ["lesson_6"], indirect=True)
+def test_dbt_partitioned_incremental_assets(setup_dbt_env): # noqa: F811
+    from dagster_and_dbt.lesson_6.assets import dbt
+    from dagster_and_dbt.lesson_6.resources import dbt_resource
+
+    dbt_analytics_assets = dg.load_assets_from_modules(modules=[dbt])
+
     result = dg.materialize(
         assets=[*dbt_analytics_assets],
         resources={
@@ -17,5 +21,8 @@ def test_dbt_partitioned_incremental_assets():
     assert result.success
 
 
-def test_def_can_load():
+@pytest.mark.parametrize("setup_dbt_env", ["lesson_6"], indirect=True)
+def test_def_can_load(setup_dbt_env): # noqa: F811
+    from dagster_and_dbt.lesson_6.definitions import defs
+
     assert defs
