+++ dagster_university/dagster_and_dbt/dagster_and_dbt_tests/test_lesson_7.py
@@ -1,12 +1,14 @@
 import dagster as dg
-from dagster_and_dbt.lesson_7.assets import dbt, metrics, trips
-from dagster_and_dbt.lesson_7.definitions import defs
-from dagster_and_dbt.lesson_7.resources import database_resource, dbt_resource
+import pytest
 
-dbt_analytics_assets = dg.load_assets_from_modules(modules=[dbt])
+from dagster_and_dbt_tests.fixtures import setup_dbt_env  # noqa: F401
 
 
-def test_trips_partitioned_assets():
+@pytest.mark.parametrize("setup_dbt_env", ["lesson_7"], indirect=True)
+def test_trips_partitioned_assets(setup_dbt_env): # noqa: F811
+    from dagster_and_dbt.lesson_7.assets import metrics, trips
+    from dagster_and_dbt.lesson_7.resources import database_resource
+
     assets = [
         trips.taxi_trips_file,
         trips.taxi_zones_file,
@@ -26,7 +28,11 @@ def test_trips_partitioned_assets():
     assert result.success
 
 
-def test_trips_by_week_partitioned_assets():
+@pytest.mark.parametrize("setup_dbt_env", ["lesson_7"], indirect=True)
+def test_trips_by_week_partitioned_assets(setup_dbt_env): # noqa: F811
+    from dagster_and_dbt.lesson_7.assets import metrics
+    from dagster_and_dbt.lesson_7.resources import database_resource
+
     assets = [
         metrics.trips_by_week,
     ]
@@ -40,7 +46,13 @@ def test_trips_by_week_partitioned_assets():
     assert result.success
 
 
-def test_dbt_partitioned_incremental_assets():
+@pytest.mark.parametrize("setup_dbt_env", ["lesson_7"], indirect=True)
+def test_dbt_partitioned_incremental_assets(setup_dbt_env): # noqa: F811
+    from dagster_and_dbt.lesson_7.assets import dbt
+    from dagster_and_dbt.lesson_7.resources import dbt_resource
+
+    dbt_analytics_assets = dg.load_assets_from_modules(modules=[dbt])
+
     result = dg.materialize(
         assets=[*dbt_analytics_assets],
         resources={
@@ -51,5 +63,8 @@ def test_dbt_partitioned_incremental_assets():
     assert result.success
 
 
-def test_def_can_load():
+@pytest.mark.parametrize("setup_dbt_env", ["lesson_7"], indirect=True)
+def test_def_can_load(setup_dbt_env): # noqa: F811
+    from dagster_and_dbt.lesson_7.definitions import defs
+
     assert defs
